- extracted logic from AbstractStepAdapter to the sample adapter - refactored step adapters in the sample app
