+++ sample/src/main/java/com/stepstone/stepper/sample/MainActivity.java
@@ -21,8 +21,6 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import com.stepstone.stepper.sample.nofrag.NoFragActivity;
-
 import butterknife.ButterKnife;
 import butterknife.OnClick;
 
@@ -95,9 +93,9 @@ public class MainActivity extends AppCompatActivity {
         startActivity(new Intent(this, ReturnButtonActivity.class));
     }
 
-    @OnClick(R.id.noFrag)
+    @OnClick(R.id.noFragments)
     public void onNoFrag(View view){
-        startActivity(new Intent(this, NoFragActivity.class));
+        startActivity(new Intent(this, NoFragmentsActivity.class));
     }
 
 }
