--- sample/src/main/java/com/stepstone/stepper/sample/nofrag/NoFragAdapter.java
@@ -1,29 +1,0 @@
-package com.stepstone.stepper.sample.nofrag;
-
-import android.content.Context;
-
-import com.stepstone.stepper.adapter.AbstractStepAdapter;
-
-/**
- * Created by leonardo on 18/12/16.
- */
-
-public class NoFragAdapter extends AbstractStepAdapter {
-
-    private final Context context;
-
-    public NoFragAdapter(Context context) {
-        this.context = context;
-    }
-
-    @Override
-    public NoFragView createStep(int position) {
-        return new NoFragView(context);
-    }
-
-    @Override
-    public int getCount() {
-        return 3;
-    }
-
-}
