+++ sample/src/main/java/com/stepstone/stepper/sample/step/view/StepViewSample.java
@@ -1,12 +1,9 @@
-package com.stepstone.stepper.sample.nofrag;
+package com.stepstone.stepper.sample.step.view;
 
-import android.annotation.TargetApi;
 import android.content.Context;
-import android.os.Build;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.text.Html;
-import android.util.AttributeSet;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.animation.AnimationUtils;
@@ -22,7 +19,7 @@ import com.stepstone.stepper.sample.R;
  * Created by leonardo on 18/12/16.
  */
 
-public class NoFragView extends FrameLayout implements Step {
+public class StepViewSample extends FrameLayout implements Step {
 
     private static final int TAP_THRESHOLD = 2;
 
@@ -33,24 +30,8 @@ public class NoFragView extends FrameLayout implements Step {
 
     private Button button;
 
-    public NoFragView(Context context) {
+    public StepViewSample(Context context) {
         super(context);
-        init(context);
-    }
-
-    public NoFragView(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        init(context);
-    }
-
-    public NoFragView(Context context, AttributeSet attrs, int defStyleAttr) {
-        super(context, attrs, defStyleAttr);
-        init(context);
-    }
-
-    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-    public NoFragView(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
-        super(context, attrs, defStyleAttr, defStyleRes);
         init(context);
     }
 
@@ -58,7 +39,7 @@ public class NoFragView extends FrameLayout implements Step {
     protected void onAttachedToWindow() {
         super.onAttachedToWindow();
         Context c = getContext();
-        if(c instanceof OnNavigationBarListener){
+        if (c instanceof OnNavigationBarListener) {
             this.onNavigationBarListener = (OnNavigationBarListener) c;
         }
     }
@@ -69,7 +50,7 @@ public class NoFragView extends FrameLayout implements Step {
         this.onNavigationBarListener = null;
     }
 
-    private void init(Context context){
+    private void init(Context context) {
         View v = LayoutInflater.from(context).inflate(R.layout.fragment_step, this, true);
         button = (Button) v.findViewById(R.id.button);
