+++ sample/src/main/res/layout/activity_main.xml
@@ -89,10 +89,10 @@
             android:text="@string/show_back_button" />
 
         <Button
-            android:id="@+id/noFrag"
+            android:id="@+id/noFragments"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/no_frags" />
+            android:text="@string/no_fragments" />
 
     </LinearLayout>
 </ScrollView>
